requirements.txt
# RFLPA Requirements
# Python 3.11+ recommended
# Install: pip install -r requirements.txt

# Core machine learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0

# Data processing and visualization
pandas>=1.3.0
matplotlib>=3.3.0

# Cryptographic operations (for secure aggregation)
pycryptodome>=3.15.0

# High-precision arithmetic (for finite field operations)
gmpy2>=2.1.0

# Optional: CUDA support
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/
